+++ test/fixtures/glide-with-hg-dep/main.go
@@ -1,9 +1,0 @@
-package main
-
-import "fmt"
-import "bitbucket.org/pkg/urlenc"
-
-func main() {
-	i := urlenc.Unmarshal
-	fmt.Println(i)
-}